One Front Door/SourceCode/MerchAngular2ReleaseiPhone/src/app/common/menu/user-dropdown/user-dropdown.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { UserDropdownComponent } from './user-dropdown.component';

describe('Component: UserDropdown', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [UserDropdownComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([UserDropdownComponent],
      (component: UserDropdownComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(UserDropdownComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(UserDropdownComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-user-dropdown></app-user-dropdown>
  `,
  directives: [UserDropdownComponent]
})
class UserDropdownComponentTestController {
}

